backend/requirements.txt
flask
flask-session
flask-cors
spotipy
requests
textblob
nltk
pandas
scikit-learn
python-dotenv

